Fixed Player Bullet Issues, Applied New Package
Fixed an issue with the Bullet where due to it not being labelled as a Trigger it had caused the enemy ships to not be able to be destroyed and do some funky things to the Machine gun and triple shot weapons. these and a pickup item are part of a package i applied to this project (a seperate version of the Task 1B Repository, split for marking purposes).