Made a Triple Shot Boss and Removed Scripts
Utilised code changes to allow and make a enemy Boss Ship, this was done with changing the WeaponTripleShot Script to utilise the Tags that were used to destroy the game objects to check the bullet type and change the direction of the bullets in the script depending on the Tag on them. this is how I found out about this function: https://discussions.unity.com/t/referencing-multiple-tags-in-if-statement/559786
secondly I Removed 2 now redundant scripts. the original Player Movement Script and Shooting Script.