Removed Scripts and Applied them to PlayerMovement
I have Removed most of the Movement Based Scripts (EnemyMovement, EntityMoveForward, and MoveConstantly) and moved their functionality to PlayerMovement to Clean up Space, I was unsure what to do with the Engine Script (Most likely This but it was too late by now to ask), so I have Kept it under the understanding that it would either be Deleted or had Functions moved to there. how I got the MoveConstantly Script into the PlayerMovement without the 2 Clashing or causing unnecessary usage of the Update Function was with a command called "enabled = false" which I found on this website: https://discussions.unity.com/t/what-is-the-command-to-stop-the-update-function-ingame-c/177354

Did attempt to Get the Rotate Script into the PlayerMovement Script however when I tried there was an Error with the "Random" part of it. so I have left the Rotate Script Separate to the Rest of the Movement Functions.